fix(core): eliminate strict compilation warnings across GCC/Clang/MSVC
Add explicit type casts for size_t/size_type conversions, fix memset int param, remove EMH_UNLIKELY in MSVC-sensitive comparisons. Zero warnings under -Wconversion and /W4.

+++ include/emhash/hash_set8.hpp
@@ -431,7 +431,7 @@ public:
     allocator_type get_allocator() const { return allocator_type(_pair_allocator); }
 
     void max_load_factor(float mlf) {
-        if (mlf < 0.992 && mlf > EMH_MIN_LOAD_FACTOR) {
+        if (mlf < 0.992f && mlf > EMH_MIN_LOAD_FACTOR) {
             _mlf = (uint32_t)((1 << 27) / mlf);
             if (_num_buckets > 0)
                 rehash(_num_buckets);
@@ -927,7 +927,7 @@ public:
 
     void rebuild(size_type num_buckets, size_type required_buckets, size_type old_num_buckets) noexcept {
         dealloc_index(_index, old_num_buckets);
-        const auto need_size = std::max((size_type)((double)num_buckets * max_load_factor()) + 4, required_buckets + 2);
+        const auto need_size = std::max((size_type)((double)num_buckets * (double)max_load_factor()) + 4, required_buckets + 2);
         auto new_pairs = alloc_bucket(need_size);
         if (is_trivially_copyable()) {
             if (_pairs)
